These changes to the `shoebox-helm` chart appear to conditionally enable mounting a volume named `config-volume` at `/app/config` within the deployed container.
Previously, the volume mount was unconditionally included in the deployment template.  This update moves the `volumeMounts` section under a conditional block controlled by the `Values.volumeMounts` setting.  If this setting is defined (and evaluates to a truthy value in the template rendering), then the `config-volume` mount, along with any other mounts defined in `Values.volumeMounts`, will be included in the deployment. If `Values.volumeMounts` isn't defined or is falsy, the `config-volume` mount will be omitted.

The `Chart.yaml` file is also updated, bumping the chart version from 0.0.4 to 0.0.5, reflecting this change.  This is standard practice for Helm charts to track changes.

In essence, this change provides more flexibility in deployments. Users can now choose whether or not to mount the `config-volume` based on their needs by setting the `Values.volumeMounts` parameter during helm install or upgrade.  This is useful for scenarios where configuration might be provided through other means, like ConfigMaps or environment variables, and mounting a specific volume is not always desired.